+++ extra/scripts/phase-3.ps1
@@ -2,17 +2,17 @@
 # Uncomment next line to skip this
 
 # exit 0
-Write-Output "Start of Phase-3"
-    $puppet_version="5.5.16"
+Write-Output "Start of Phase 3"
+    $puppet_version="5.5.18"
     # add your stuff here
     # add my myget source
     try {
         choco source add -n=public -s"https://www.myget.org/F/public-choco" --priority=10
         # install puppet/stop/clean
-        choco install puppet-agent --version $puppet_version -ia '"PUPPET_MASTER_SERVER=foreman.spcph.local"' -y
+        choco install puppet-agent --version $puppet_version -ia '"PUPPET_MASTER_SERVER=foreman.spcph.local"' -y --no-progress
     }
     catch {
-        Write-Output "Phase-3 - Install source or install puppet-agent failed"
+        Write-Output "Phase 3 - Install source or install puppet-agent failed"
     }
     try {
         Get-Service -name puppet -ErrorAction Silentlycontinue|Stop-Service -Force -ErrorAction Silentlycontinue
@@ -21,7 +21,7 @@ Write-Output "Start of Phase-3"
         Remove-Item "c:\ProgramData\PuppetLabs\puppet\etc\ssl\" -Recurse -Force -ErrorAction Silentlycontinue
     }
     catch {
-        Write-Output "Phase-3 - Cleaning or stopping "
+        Write-Output "Phase 3 - Cleaning or stopping "
     }
-Write-Output "End of Phase-3"
+Write-Output "End of Phase 3"
 exit 0